docs(table): remove old stories
