+++ packages/react/src/components/PageTitleBar/PageTitleBar.story.jsx
@@ -10,13 +10,14 @@ import TileCatalogNew from '../TileCatalogNew/TileCatalogNew';
 import { getTiles } from '../TileCatalogNew/TileCatalogNew.story';
 import Button from '../Button';
 import StatefulTable from '../Table/StatefulTable';
-import { initialState } from '../Table/Table.story';
 import { settings } from '../../constants/Settings';
+import { getInitialState } from '../Table/Table.story.helpers';
 
 import PageTitleBar from './PageTitleBar';
 import PageTitleBarREADME from './PageTitleBar.mdx';
 
 const { prefix } = settings;
+const initialTableState = getInitialState();
 
 export const commonPageTitleBarProps = {
   title: 'Page title',
@@ -99,8 +100,8 @@ export const Base = () => {
         forceContentOutside={boolean('forceContentOutside', false)}
         content={
           <div style={{ display: 'flex', alignItems: 'center' }}>
-            <StatefulTable {...initialState} data={initialState.data.slice(0, 5)} />
-            <StatefulTable {...initialState} data={initialState.data.slice(0, 5)} />
+            <StatefulTable {...initialTableState} data={initialTableState.data.slice(0, 5)} />
+            <StatefulTable {...initialTableState} data={initialTableState.data.slice(0, 5)} />
           </div>
         }
         stackBreadcrumbsWithTabs={boolean('stackBreadcrumbsWithTabs', false)}
